Functions are not equivalent
