+++ pyccel/codegen/printing/ccode.py
@@ -1076,6 +1076,57 @@ class CCodePrinter(CodePrinter):
         """convert dotted Variable to their C equivalent"""
         return '{}.{}'.format(self._print(expr.lhs), self._print(expr.name))
 
+    @staticmethod
+    def _new_slice_with_processed_arguments(_slice, array_size, allow_negative_index):
+        """ Create new slice with informations collected from old slice and decorators
+
+        Parameters
+        ----------
+            _slice : Slice
+                slice needed to collect (start, stop, step)
+            array_size : PyccelArraySize
+                call to function size()
+            allow_negative_index : Bool
+                True when the decorator allow_negative_index is present
+        Returns
+        -------
+            Slice
+        """
+        start = LiteralInteger(0) if _slice.start is None else _slice.start
+        stop = array_size if _slice.stop is None else _slice.stop
+
+        # negative start and end in slice
+        if isinstance(start, PyccelUnarySub) and isinstance(start.args[0], LiteralInteger):
+            start = PyccelMinus(array_size, start.args[0], simplify = True)
+        elif allow_negative_index and not isinstance(start, (LiteralInteger, PyccelArraySize)):
+            start = IfTernaryOperator(PyccelLt(start, LiteralInteger(0)),
+                            PyccelMinus(array_size, start, simplify = True), start)
+
+        if isinstance(stop, PyccelUnarySub) and isinstance(stop.args[0], LiteralInteger):
+            stop = PyccelMinus(array_size, stop.args[0], simplify = True)
+        elif allow_negative_index and not isinstance(stop, (LiteralInteger, PyccelArraySize)):
+            stop = IfTernaryOperator(PyccelLt(stop, LiteralInteger(0)),
+                            PyccelMinus(array_size, stop, simplify = True), stop)
+
+        # steps in slices
+        step = _slice.step
+
+        if step is None:
+            step = LiteralInteger(1)
+
+        # negative step in slice
+        elif isinstance(step, PyccelUnarySub) and isinstance(step.args[0], LiteralInteger):
+            start = PyccelMinus(array_size, LiteralInteger(1), simplify = True) if _slice.start is None else start
+            stop = LiteralInteger(0) if _slice.stop is None else stop
+
+        # variable step in slice
+        elif allow_negative_index and step and not isinstance(step, LiteralInteger):
+            og_start = start
+            start = IfTernaryOperator(PyccelGt(step, LiteralInteger(0)), start, PyccelMinus(stop, LiteralInteger(1), simplify = True))
+            stop = IfTernaryOperator(PyccelGt(step, LiteralInteger(0)), stop, og_start)
+
+        return Slice(start, stop, step)
+
     def _print_PyccelArraySize(self, expr):
         return '{}.shape[{}]'.format(expr.arg, expr.index)
