+++ pyccel/codegen/printing/fcode.py
@@ -2625,6 +2625,62 @@ class FCodePrinter(CodePrinter):
 
         return "%s(%s)" % (base_code, ", ".join(inds))
 
+    @staticmethod
+    def _new_slice_with_processed_arguments(_slice, array_size, allow_negative_index):
+        """ Create new slice with informations collected from old slice and decorators
+
+        Parameters
+        ----------
+            _slice : Slice
+                slice needed to collect (start, stop, step)
+            array_size : PyccelArraySize
+                call to function size()
+            allow_negative_index : Bool
+                True when the decorator allow_negative_index is present
+        Returns
+        -------
+            Slice
+        """
+        start = _slice.start
+        stop = _slice.stop
+        step = _slice.step
+
+        # negative start and end in slice
+        if isinstance(start, PyccelUnarySub) and isinstance(start.args[0], LiteralInteger):
+            start = PyccelMinus(array_size, start.args[0], simplify = True)
+        elif start is not None and allow_negative_index and not isinstance(start,LiteralInteger):
+            start = IfTernaryOperator(PyccelLt(start, LiteralInteger(0)),
+                        PyccelAdd(array_size, start, simplify = True), start)
+
+        if isinstance(stop, PyccelUnarySub) and isinstance(stop.args[0], LiteralInteger):
+            stop = PyccelMinus(array_size, stop.args[0], simplify = True)
+        elif stop is not None and allow_negative_index and not isinstance(stop, LiteralInteger):
+            stop = IfTernaryOperator(PyccelLt(stop, LiteralInteger(0)),
+                        PyccelAdd(array_size, stop, simplify = True), stop)
+
+        # negative step in slice
+        if isinstance(step, PyccelUnarySub) and isinstance(step.args[0], LiteralInteger):
+            stop = PyccelAdd(stop, LiteralInteger(1), simplify = True) if stop is not None else LiteralInteger(0)
+            start = start if start is not None else PyccelMinus(array_size, LiteralInteger(1), simplify = True)
+
+        # variable step in slice
+        elif step and allow_negative_index and not isinstance(step, LiteralInteger):
+            if start is None :
+                start = IfTernaryOperator(PyccelGt(step, LiteralInteger(0)),
+                    LiteralInteger(0), PyccelMinus(array_size , LiteralInteger(1), simplify = True))
+
+            if stop is None :
+                stop = IfTernaryOperator(PyccelGt(step, LiteralInteger(0)),
+                    PyccelMinus(array_size, LiteralInteger(1), simplify = True), LiteralInteger(0))
+            else :
+                stop = IfTernaryOperator(PyccelGt(step, LiteralInteger(0)),
+                    stop, PyccelAdd(stop, LiteralInteger(1), simplify = True))
+
+        elif stop is not None:
+            stop = PyccelMinus(stop, LiteralInteger(1), simplify = True)
+
+        return Slice(start, stop, step)
+
     def _print_Slice(self, expr):
         if expr.start is None or  isinstance(expr.start, Nil):
             start = ''
@@ -2633,7 +2689,7 @@ class FCodePrinter(CodePrinter):
         if (expr.stop is None) or isinstance(expr.stop, Nil):
             stop = ''
         else:
-            stop = self._print(PyccelMinus(expr.stop, LiteralInteger(1), simplify = True))
+            stop = self._print(expr.stop)
         if expr.step is not None :
             return '{0}:{1}:{2}'.format(start, stop, self._print(expr.step))
         return '{0}:{1}'.format(start, stop)
